requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.0
playwright>=1.40.0
pyyaml>=6.0
requests>=2.31.0
